+++ src/utils.ts
@@ -30,6 +30,9 @@ export default class Utils {
       darkPalette: workspaceConfiguration.get<string>("darkPalette"),
       lightPalette: workspaceConfiguration.get<string>("lightPalette"),
       colorfulSyntax: workspaceConfiguration.get<boolean>("colorfulSyntax"),
+      semanticHighlighting: workspaceConfiguration.get<boolean>(
+        "semanticHighlighting"
+      ),
       italicKeywords: workspaceConfiguration.get<boolean>("italicKeywords"),
       italicComments: workspaceConfiguration.get<boolean>("italicComments"),
     };
@@ -40,12 +43,14 @@ export default class Utils {
       dark: {
         name: "Gruvbox Material Dark",
         type: "dark",
+        semanticHighlighting: configuration.semanticHighlighting,
         colors: getWorkbench(configuration, "dark"),
         tokenColors: getSyntax(configuration, "dark"),
       },
       light: {
         name: "Gruvbox Material Light",
         type: "light",
+        semanticHighlighting: configuration.semanticHighlighting,
         colors: getWorkbench(configuration, "light"),
         tokenColors: getSyntax(configuration, "light"),
       },
@@ -89,6 +94,7 @@ export default class Utils {
     // {{{
     return (
       configuration.colorfulSyntax === false &&
+      configuration.semanticHighlighting === true &&
       configuration.italicKeywords === false &&
       configuration.italicComments === true &&
       configuration.lightPalette === "material" &&